refactor(worker): reorganize tracker imports and enhance worker initialization
- Updated import paths for `TileTracker` and related modules to improve project structure and clarity.
- Introduced a new `worker-tracker-registry.ts` file to centralize tracker initialization and configuration management.
- Refactored `chess-worker.ts` to utilize the new registry for loading custom trackers and managing analysis configurations.
- Enhanced the `GameParser` class to streamline SAN move processing by delegating responsibilities to `SanApplier` and `SanParser` classes.
- Improved documentation and code organization for better maintainability and readability.
