+++ src/index.ts
@@ -2,6 +2,6 @@ import Chessalyzer from './core/chessalyzer';
 import BaseTracker from './tracker/base-tracker';
 import GameTracker from './tracker/game-tracker-base';
 import PieceTracker from './tracker/piece-tracker-base';
-import TileTracker from './tracker/tile-tracker-base';
+import TileTracker from './tracker/tile/tile-tracker-base';
 
 export { Chessalyzer, PieceTracker, GameTracker, TileTracker, BaseTracker };
